Shifted agent to its own class from structure
