+++ include/lazy_traffic_agent.hpp
@@ -1,0 +1,44 @@
+// Created by Indraneel on 22/09/22
+
+#ifndef LAZY_TRAFFIC_AGENT_H
+#define LAZY_TRAFFIC_AGENT_H
+
+#include <string>
+#include <queue>
+
+// ROS stuff
+#include <geometry_msgs/Twist.h>
+#include <geometry_msgs/PoseStamped.h>
+#include <geometry_msgs/TransformStamped.h>
+#include <ros/ros.h>
+
+class Agent {
+
+public:
+    Agent() : name_(""), robot_frame_id_(""), current_path_(), current_pose_() {}
+    Agent(std::string name, ros::NodeHandle nh) : name_(name), robot_frame_id_(name + "/base_link"), nh_(nh) {
+        // Initialise publisher
+        pub_vel_ = nh_.advertise<geometry_msgs::Twist>("/robosar_agent_bringup_node/" + name + "/cmd_vel", 1);
+    }
+    ~Agent() {}
+
+    void stop_agent(void) {
+        // TODO Check if velocity is non zero 
+        geometry_msgs::Twist vel;
+        vel.linear.x = 0.0;
+        vel.angular.z = 0.0;
+        pub_vel_.publish(vel);
+    }
+
+    std::string robot_frame_id_;
+    std::queue<geometry_msgs::PoseStamped> current_path_;
+    geometry_msgs::TransformStamped current_pose_;
+private:
+    ros::Publisher pub_vel_;
+    std::string name_;
+    ros::NodeHandle nh_;
+
+
+};
+
+#endif // LAZY_TRAFFIC_AGENT_H
