+++ include/lazy_traffic_controller.hpp
@@ -10,12 +10,12 @@
 #include <mutex>
 
 #include "robosar_messages/robosar_controller.h"
+#include "lazy_traffic_agent.hpp"
 
 // ROS stuff
 #include <tf/tf.h>
 #include <tf2_ros/transform_listener.h>
 #include <ros/console.h>
-#include <ros/ros.h>
 #include <std_msgs/Bool.h>
 #include <geometry_msgs/Twist.h>
 
@@ -37,25 +37,8 @@ private:
     bool controllerServiceCallback(robosar_messages::robosar_controller::Request &req,
                                    robosar_messages::robosar_controller::Response &res);
     void updateAgentPoses();
-    
+    void updatePreferredVelocities();
 
-    typedef struct Agent {
-        std::string name;
-        ros::Publisher pub_vel_;
-        std::string robot_frame_id_;
-        std::queue<geometry_msgs::PoseStamped> current_path;
-        geometry_msgs::TransformStamped current_pose;
-
-        void stop_agent(void) {
-            // TODO Check if velocity is non zero 
-            geometry_msgs::Twist vel;
-            vel.linear.x = 0.0;
-            vel.angular.z = 0.0;
-            pub_vel_.publish(vel);
-        }
-
-    } Agent_s;
-    
 
     // miscellanous
     std::thread traffic_controller_thread_;
@@ -66,7 +49,7 @@ private:
     std::mutex map_mutex;
 
     // controller data structures
-    std::map<std::string, Agent_s> agent_map_;
+    std::map<std::string, Agent> agent_map_;
     std::set<std::string> active_agents;
 
     // ROS stuff
